Decompose PodmanBackend and PodmanProxyManager god-objects
Both classes exceeded coding standards (400-line file limit, 50-line
method limit, 20-method class limit). Extract collaborating classes
while preserving the public API (Backend protocol, __init__.py exports).

PodmanBackend: 729 lines / 38 methods → 402 lines / 20 public methods
PodmanProxyManager: 641 lines → 400 lines

New files:
- session_setup.py: SessionSetup class (credential injection, config
  sync, port URLs, container startup orchestration)
- ca_cert.py: CACertDistributor (CA cert distribution logic)
- proxy_credentials.py: ProxyCredentialManager (secret management)

Key changes:
- Moved session validation helpers to helpers.py as free functions
- Deleted dead _recreate_proxy method (-58 lines)
- Decomposed create_session from 143 lines to ~37 lines
- Inlined _start_port_forward into its 2 callers
- Updated KNOWN_ISSUES.md REFACTOR-003 to reflect resolved items
