+++ tests/test_docker_compat.py
@@ -238,7 +238,7 @@ class TestDockerVolumePermissions:
         mock_run.return_value = MagicMock(returncode=0, stdout="", stderr="")
         engine = ContainerEngine("docker")
         backend = PodmanBackend(engine=engine)
-        backend._fix_volume_permissions("paude-test")
+        backend._setup.fix_volume_permissions("paude-test")
         # Find the chown call
         chown_calls = [c for c in mock_run.call_args_list if "chown" in c[0][0]]
         assert len(chown_calls) == 1
@@ -259,7 +259,7 @@ class TestDockerVolumePermissions:
         engine = ContainerEngine("podman")
         backend = PodmanBackend(engine=engine)
         # Should be a no-op, no subprocess call needed
-        backend._fix_volume_permissions("paude-test")
+        backend._setup.fix_volume_permissions("paude-test")
 
 
 class TestDockerCredentialInjection:
@@ -277,7 +277,7 @@ class TestDockerCredentialInjection:
 
         engine = ContainerEngine("docker")
         backend = PodmanBackend(engine=engine)
-        backend._inject_stub_credentials("paude-test")
+        backend._setup.inject_stub_credentials("paude-test")
 
         exec_calls = [
             c
@@ -338,7 +338,7 @@ class TestDockerCredentialInjection:
 
         engine = ContainerEngine("docker")
         backend = PodmanBackend(engine=engine)
-        backend._inject_stub_credentials("paude-test")
+        backend._setup.inject_stub_credentials("paude-test")
 
         exec_calls = [
             c
@@ -362,7 +362,7 @@ class TestStubCredentialInjection:
 
         engine = ContainerEngine("podman")
         backend = PodmanBackend(engine=engine)
-        backend._inject_stub_credentials("paude-test")
+        backend._setup.inject_stub_credentials("paude-test")
 
         exec_calls = [
             c
@@ -382,7 +382,7 @@ class TestCodexSyntheticAuth:
         mock_run.return_value = MagicMock(returncode=0, stdout="", stderr="")
         backend = PodmanBackend(engine=ContainerEngine("podman"))
 
-        backend._inject_codex_auth("paude-test", chatgpt_mode=True)
+        backend._setup.inject_codex_auth("paude-test", chatgpt_mode=True)
 
         injections = [
             c for c in mock_run.call_args_list if c[1].get("input") is not None
@@ -396,7 +396,7 @@ class TestCodexSyntheticAuth:
         mock_run.return_value = MagicMock(returncode=0, stdout="", stderr="")
         backend = PodmanBackend(engine=ContainerEngine("podman"))
 
-        backend._inject_codex_auth("paude-test", chatgpt_mode=False)
+        backend._setup.inject_codex_auth("paude-test", chatgpt_mode=False)
 
         exec_args = [c[0][0] for c in mock_run.call_args_list if "exec" in c[0][0]]
         assert any(
